Kumar Forging LTD/requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
psycopg2-binary>=2.9.5
numpy>=1.21.0
matplotlib>=3.5.0
umap-learn>=0.5.3
google-generativeai>=0.3.0
ragas>=0.0.16
datasets>=2.12.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2 